Add Barrier Script and Bonus

+++ Assets/Scripts/Weapon.cs
@@ -7,7 +7,7 @@ public class Weapon : ScriptableObject
 {
     public GameObject Bulletprefab;
     public float speed;
-    public float Damage = 10;
+    public int Damage = 10;
